Everything to int64

+++ pkg/proxmox/node_models.go
@@ -6,16 +6,16 @@ type NodeResponse struct {
 
 type Node struct {
 	Type           string  `json:"type"`
-	Maxcpu         int     `json:"maxcpu"`
+	Maxcpu         int64   `json:"maxcpu"`
 	Cpu            float64 `json:"cpu"`
 	Status         string  `json:"status"`
-	Maxmem         int     `json:"maxmem"`
+	Maxmem         int64   `json:"maxmem"`
 	SslFingerprint string  `json:"ssl_fingerprint"`
-	Mem            int     `json:"mem"`
+	Mem            int64   `json:"mem"`
 	Id             string  `json:"id"`
 	Node           string  `json:"node"`
 	Disk           int64   `json:"disk"`
-	Uptime         int     `json:"uptime"`
+	Uptime         int64   `json:"uptime"`
 	Maxdisk        int64   `json:"maxdisk"`
 	Level          string  `json:"level"`
 }
